+++ src/helpers/keyCodes.js
@@ -2,7 +2,6 @@ const KEY_CODES = {
   Space: 32,
   Enter: 13,
   Esc: 27,
-  Tab: 8,
   Left: 37,
   Up: 38,
   Right: 39,